Introducing code coverage and first approaches to reduce cyc complexity
