+++ .gitignore
@@ -1,4 +1,5 @@
 .env
 node_modules
+coverage
 dist
 .idea
