+++ package.json
@@ -9,7 +9,7 @@
   "scripts": {
     "start": "./start_locally.sh",
     "build": "rimraf dist && ncc build src/app.js -o dist",
-    "test": "jest",
+    "test": "jest --coverage=true --coverageReporter=json,lcov,text",
     "ci": "start-server-and-test http://127.0.0.1:3000",
     "deploy": "cd docker && docker build .. -f Dockerfile --no-cache -t restaurant",
     "docker": "./run_docker.sh",
